requirements.txt
git+https://github.com/openai/whisper.git
sounddevice
asyncio
numpy
simplewebsocketserver
